feat: manually declare entry points to skip scanning
