+++ package.json
@@ -21,9 +21,9 @@
   "bugs": "https://github.com/microsoft/react-native-lazy-index/issues",
   "license": "MIT",
   "files": [
-    "*.js"
+    "src/*.js"
   ],
-  "main": "index.js",
+  "main": "src/index.js",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/microsoft/react-native-lazy-index.git"
