+++ test/__fixtures__/AppRegistry/package.json
@@ -1,6 +1,6 @@
 {
-  "experiences": {
-    "Component-01": "./component-01",
-    "Component-02": "./component-02"
-  }
+  "experiences": [
+    "./component-01",
+    "./component-02"
+  ]
 }
