+++ test/__fixtures__/BatchedBridge/package.json
@@ -1,6 +1,6 @@
 {
-  "experiences": {
-    "Module-01": "./module-01",
-    "Module-02": "./module-02"
-  }
+  "experiences": [
+    "./module-01",
+    "./module-02"
+  ]
 }
